# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: Financial snapshot
  - name: goals
    description: User goals
  - name: roadmap
    description: Personalized roadmap steps
  - name: documents
    description: Document filing system
  - name: opportunities
    description: Lending and investment marketplace
  - name: dashboard
    description: Dashboard aggregates and readiness scores
  - name: openai
    description: Pepper AI chat and voice operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the financial snapshot
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update the financial snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /roadmap:
    get:
      operationId: listRoadmapSteps
      tags: [roadmap]
      summary: List roadmap steps
      responses:
        "200":
          description: Roadmap steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoadmapStep"
    post:
      operationId: createRoadmapStep
      tags: [roadmap]
      summary: Create a roadmap step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapInput"
      responses:
        "201":
          description: Created step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapStep"
  /roadmap/{id}:
    patch:
      operationId: updateRoadmapStep
      tags: [roadmap]
      summary: Update a roadmap step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapUpdate"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapStep"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document (move bucket / reorder / rename)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List lending and investment opportunities
      responses:
        "200":
          description: Opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"

  /scores:
    get:
      operationId: getReadinessScores
      tags: [dashboard]
      summary: Get readiness scores computed from the financial snapshot
      responses:
        "200":
          description: Scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadinessScore"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Aggregated dashboard summary
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /openai/conversations:
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /openai/conversations/{id}/voice-messages:
    post:
      operationId: sendOpenaiVoiceMessage
      tags: [openai]
      summary: Send audio and receive a streaming transcript and voice response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiVoiceMessageInput"
      responses:
        "200":
          description: SSE stream of transcript and audio chunks
          content:
            text/event-stream: {}
  /openai/transcribe:
    post:
      operationId: transcribeOpenaiAudio
      tags: [openai]
      summary: Transcribe audio to text (dictation, no AI reply)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiTranscribeInput"
      responses:
        "200":
          description: Transcribed text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiTranscribeOutput"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Profile:
      type: object
      properties:
        id:
          type: integer
        displayName:
          type: string
        monthlyIncome:
          type: integer
        monthlyExpenses:
          type: integer
        cashSavings:
          type: integer
        otherAssets:
          type: integer
        totalDebt:
          type: integer
        creditScore:
          type: integer
        preferredVoice:
          type: string
        onboarded:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - displayName
        - monthlyIncome
        - monthlyExpenses
        - cashSavings
        - otherAssets
        - totalDebt
        - creditScore
        - preferredVoice
        - onboarded
        - updatedAt
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        monthlyIncome:
          type: integer
        monthlyExpenses:
          type: integer
        cashSavings:
          type: integer
        otherAssets:
          type: integer
        totalDebt:
          type: integer
        creditScore:
          type: integer
        preferredVoice:
          type: string
          enum: ["female", "male"]
        onboarded:
          type: boolean

    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        targetAmount:
          type: integer
        currentAmount:
          type: integer
        targetDate:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: integer
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - targetAmount
        - currentAmount
        - status
        - priority
        - createdAt
    GoalInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        category:
          type: string
          enum: ["homeownership", "investing", "passive_income", "credit", "debt", "wealth"]
        targetAmount:
          type: integer
        currentAmount:
          type: integer
        targetDate:
          type: string
        status:
          type: string
          enum: ["active", "achieved", "paused"]
        priority:
          type: integer
        note:
          type: string
      required:
        - title
        - category
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        category:
          type: string
          enum: ["homeownership", "investing", "passive_income", "credit", "debt", "wealth"]
        targetAmount:
          type: integer
        currentAmount:
          type: integer
        targetDate:
          type: string
        status:
          type: string
          enum: ["active", "achieved", "paused"]
        priority:
          type: integer
        note:
          type: string

    RoadmapStep:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        orderIndex:
          type: integer
        actionLabel:
          type: ["string", "null"]
        goalId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - orderIndex
        - createdAt
    RoadmapInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: ["todo", "in_progress", "done"]
        orderIndex:
          type: integer
        actionLabel:
          type: string
        goalId:
          type: integer
      required:
        - title
    RoadmapUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: ["todo", "in_progress", "done"]
        orderIndex:
          type: integer
        actionLabel:
          type: string
        goalId:
          type: integer

    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        status:
          type: string
        orderIndex:
          type: integer
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - status
        - orderIndex
        - createdAt
    DocumentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: ["Income", "Assets", "Identity", "Property", "Credit", "Other"]
        status:
          type: string
          enum: ["needed", "in_progress", "complete"]
        orderIndex:
          type: integer
        note:
          type: string
      required:
        - name
    DocumentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: ["Income", "Assets", "Identity", "Property", "Credit", "Other"]
        status:
          type: string
          enum: ["needed", "in_progress", "complete"]
        orderIndex:
          type: integer
        note:
          type: string

    Opportunity:
      type: object
      properties:
        id:
          type: integer
        kind:
          type: string
        title:
          type: string
        summary:
          type: string
        detail:
          type: ["string", "null"]
        rate:
          type: ["string", "null"]
        term:
          type: ["string", "null"]
        minAmount:
          type: integer
        tag:
          type: ["string", "null"]
        recommended:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - kind
        - title
        - summary
        - minAmount
        - recommended
        - createdAt

    ReadinessScore:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        score:
          type: integer
        tier:
          type: string
        summary:
          type: string
      required:
        - key
        - label
        - score
        - tier
        - summary

    DashboardSummary:
      type: object
      properties:
        netWorth:
          type: integer
        monthlyCashflow:
          type: integer
        totalAssets:
          type: integer
        totalDebt:
          type: integer
        activeGoals:
          type: integer
        achievedGoals:
          type: integer
        avgReadiness:
          type: integer
        documentsComplete:
          type: integer
        documentsTotal:
          type: integer
        topScore:
          $ref: "#/components/schemas/ReadinessScore"
        nextStep:
          oneOf:
            - $ref: "#/components/schemas/RoadmapStep"
            - type: "null"
        recommendedOpportunities:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
      required:
        - netWorth
        - monthlyCashflow
        - totalAssets
        - totalDebt
        - activeGoals
        - achievedGoals
        - avgReadiness
        - documentsComplete
        - documentsTotal
        - recommendedOpportunities

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiVoiceMessageInput:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
        voice:
          type: string
          enum: ["female", "male"]
      required:
        - audio
    OpenaiTranscribeInput:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio
    OpenaiTranscribeOutput:
      type: object
      properties:
        text:
          type: string
      required:
        - text