# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Opportunity } from './opportunity';
import type { ReadinessScore } from './readinessScore';
import type { RoadmapStep } from './roadmapStep';

export interface DashboardSummary {
  netWorth: number;
  monthlyCashflow: number;
  totalAssets: number;
  totalDebt: number;
  activeGoals: number;
  achievedGoals: number;
  avgReadiness: number;
  documentsComplete: number;
  documentsTotal: number;
  topScore?: ReadinessScore;
  nextStep?: RoadmapStep | null;
  recommendedOpportunities: Opportunity[];
}