# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  name: string;
  category: string;
  status: string;
  orderIndex: number;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}