# lib/api-zod/src/generated/types/openaiVoiceMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiVoiceMessageInputVoice } from './openaiVoiceMessageInputVoice';

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
  voice?: OpenaiVoiceMessageInputVoice;
}